requirement.txt.txt
torch>=1.10.0
transformers>=4.25.1
dgl>=0.9.1
pennylane>=0.29.1
scikit-learn>=1.1.1
rdkit-pypi>=2022.9.5
matplotlib>=3.5.1
scipy>=1.8.1
numpy>=1.22.4
tqdm>=4.64.0
